% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoRasch.R
\docType{package}
\name{autoRasch}
\alias{autoRasch}
\title{autoRasch: A package for semi-automatic Rasch analysis}
\description{
This package helps user to do Rasch analysis
(semi-)automatically which is comparable to the standard Rasch analysis using
common statistics.
}
