% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{correl02_multidim}
\alias{correl02_multidim}
\title{Multidimensional polytomous data set with 0.2 correlation}
\format{
An object of class \code{data.frame} with 301 rows and 12 columns.
}
\usage{
data(correl02_multidim)
}
\description{
Multidimensional polytomous data set with 0.2 correlation
}
\keyword{datasets}
