% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{correl06_multidim}
\alias{correl06_multidim}
\title{Multidimensional polytomous data set with 0.6 Correlation}
\format{
An object of class \code{data.frame} with 301 rows and 12 columns.
}
\usage{
data(correl06_multidim)
}
\description{
Multidimensional polytomous data set with 0.6 Correlation
}
\keyword{datasets}
