% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGroup.R
\name{createGroup}
\alias{createGroup}
\title{Create Mapping Matrix of DIF Groups}
\usage{
createGroup(backInfo, idxUsed = NULL, contMethod = c("mean", "median"))
}
\arguments{
\item{backInfo}{A matrix of person background information (e.g., gender, country, age, etc);}

\item{idxUsed}{The column number of \code{backInfo} that is used for creating the mapping matrix.}

\item{contMethod}{The method of how to handle a continuous variable (e.g., mean, median). This parameter is passing a function used to split the variable into binary. The default is \code{mean}.}
}
\value{
A binary matrix that maps respondents to the groups that the respondents belongs to.
}
\description{
This function automatically create a mapping matrix based on the existing DIF inducing covariates.
}
