% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dicho_inh_dset}
\alias{dicho_inh_dset}
\title{Inhomogenous Dichotomous Data Set}
\format{
An object of class \code{data.frame} with 301 rows and 18 columns.
}
\usage{
data(dicho_inh_dset)
}
\description{
Data set with binary type responses containing three subsets
with different discrimination values.
}
\keyword{datasets}
