% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dicho_md_dset}
\alias{dicho_md_dset}
\title{Uncorrelated Multidimensional Dichotomous Data Set}
\format{
An object of class \code{data.frame} with 301 rows and 18 columns.
}
\usage{
data(dicho_md_dset)
}
\description{
Data set with binary type responses containing three subsets which
represent different uncorrelated dimensions.
}
\keyword{datasets}
