% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DIFdataset}
\alias{DIFdataset}
\title{The Two-subscale Inhomogenous Polytomous Dataset containing DIF items}
\format{
An object of class \code{data.frame} with 490 rows and 14 columns.
}
\usage{
data(DIFdataset)
}
\description{
The artificial data set of a polytomous responses (four categories) which
contains two subsets with different discrimination values and two DIF-items.
}
\keyword{datasets}
