% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit.R
\name{fitStats}
\alias{fitStats}
\alias{fitStats.pcmdif}
\alias{fitStats.pcm}
\alias{summary.fit}
\alias{plot.fit}
\title{Fit Statistics}
\usage{
fitStats(obj, isAlpha = TRUE)

\method{fitStats}{pcmdif}(obj, isAlpha = TRUE)

\method{fitStats}{pcm}(obj, isAlpha = TRUE)

\method{summary}{fit}(obj)

\method{plot}{fit}(obj, type = "n", plotx, ploty, xlab = NULL,
  ylab = NULL, use.name = FALSE, ...)
}
\arguments{
\item{obj}{The object which is resulted from models estimation (PCM or PCM-DIF)}

\item{isAlpha}{Boolean value. Are the discrimination parameters needed to be estimated using the correspond extension of the 2PL models (GPCM or GPCM-DIF, respectively).}
}
\value{
List of values

\item{alpha}{   A vector of estimated discrimination parameters for each items.}
\emph{i.fit}{   Item fit statistics.}
\itemize{
   \item{i.outfitMSQ}{   A vector of Outfit mean square values for each items.}
   \item{i.infitMSQ}{   A vector of Infit mean square values for each items.}
   \item{i.outfitZ}{   A vector of OutfitZ values for each items.}
   \item{i.infitZ}{   A vector of InfitZ values for each items.}
}
\emph{p.fit}{   Person fit statistics.}
\itemize{
   \item{p.outfitMSQ}{   A vector of Outfit mean square values for each persons.}
   \item{p.infitMSQ}{   A vector of Infit mean square values for each persons.}
   \item{p.outfitZ}{   A vector of OutfitZ values for each persons.}
   \item{p.infitZ}{   A vector of InfitZ values for each persons.}
}
}
\description{
Compute the fit statistics (Outfit, Infit, etc.) of the PCM dan PCM-DIF model estimation (items and persons)
}
\examples{
pcmdif_res <- pcm_dif(polydif_inh_dset, groups_map = c(rep(1,245),rep(0,245)))
fit_res <- fitStats(pcmdif_res)
summary(fit_res)
plot(fit_res, plotx = "gamma", ploty = "outfit")

}
