<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{generate_data}
\alias{generate_data}
\title{Generate the artificial dataset}
\usage{
generate_data(responseType = "multidim.nocorrel", theta = c(-3, 3),
  sdtheta = 6, ntheta = 301, beta = c(-2.5, 2.5), sdbeta = 4,
  nitem = 6, alpha = c(1), sdlambda = 1, ncat = 5, thGap = 0.8,
  ndim = 3, randtype = "uniform", corLevel = 0, dim.members = c())
}
\arguments{
\item{responseType}{The type of the dataset. The types include \code{multidim.nocorrel}, \code{multidim.withcorrel}, \code{discriminate}, \code{multidim.within}, and \code{testlets}.}

\item{theta}{A vector of the ability parameters range value, \code{c(min.theta,max.theta)}. It applies when the \code{randtype = "uniform"}.}

\item{sdtheta}{Standard deviation which is used to generate theta values using \code{\link[stats:rnorm]{stats::rnorm()}} with \code{n = ntheta}, \code{mean = 0}, and \code{sd = sdtheta}.It applies when the \code{randtype = "normal"}.}

\item{ntheta}{The number of the observations.}

\item{beta}{A vector of the item difficulty parameters range value, \code{c(min.beta,max.beta)}. It applies when the \code{randtype = "uniform"}.}

\item{sdbeta}{Standard deviation which is used to generate item location values using \code{\link[stats:rnorm]{stats::rnorm()}} with \code{n = nitem}, \code{mean = 0}, and \code{sd = sdbeta}.It applies when the \code{randtype = "normal"}.}

\item{nitem}{The number of the items in each subgroup.}

\item{alpha}{A vector of the discrimination parameters apply to each items.}

\item{sdlambda}{A vector of the standard deviation to simulate the testlet (local dependency) effect. The effect is added using \code{\link[stats:rnorm]{stats::rnorm()}} with \code{n = ntheta}, \code{mean = 0}, and \code{sd = sdlambda}}

\item{ncat}{The number of the response categories}

\item{thGap}{The difference between adjacent threshold.}

\item{ndim}{The number of subgroups (dimensions/testlets) created.}

\item{randtype}{The randomize type. This includes \code{uniform} and \code{normal}.}

\item{dim.members}{The list of item members in each dimension.}
}
\description{
This function generates simulated datasets with different attributes
}
\examples{
# Generate multidimensional dataset which having correlation of 0.2 between the dimensions
correl02_multidim <- generate_data(responseType = "multidim.withcorrel", corLevel = 0.2)

# Generate multidimensional dataset with some items relate to more than one dimension.
withinItem_multidim <- generate_data(responseType = "multidim.within", ndim = 3, dim.members = list(c(1:6,13),c(3,7:12),c(5,13:18)))

#generate dataset which consist of two bundle items with different level of local dependency effect.
testlets_dataset <- generate_data(responseType = "testlets", ndim = 2, sdlambda = c(0,4))

}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{generate_data}
\alias{generate_data}
\title{Generate the artificial dataset}
\usage{
generate_data(responseType = "multidim.nocorrel", theta = c(-3, 3),
  sdtheta = 6, ntheta = 301, beta = c(-2.5, 2.5), sdbeta = 4,
  nitem = 6, alpha = c(1), sdlambda = 1, ncat = 5, thGap = 0.8,
  ndim = 3, randtype = "uniform", corLevel = 0, dim.members = c())
}
\arguments{
\item{responseType}{The type of the dataset. The types include \code{multidim.nocorrel}, \code{multidim.withcorrel}, \code{discriminate}, \code{multidim.within}, and \code{testlets}.}

\item{theta}{A vector of the ability parameters range value, \code{c(min.theta,max.theta)}. It applies when the \code{randtype = "uniform"}.}

\item{sdtheta}{Standard deviation which is used to generate theta values using \code{\link[stats:list]{rnorm}} with \code{n = ntheta}, \code{mean = 0}, and \code{sd = sdtheta}.It applies when the \code{randtype = "normal"}.}

\item{ntheta}{The number of the observations.}

\item{beta}{A vector of the item difficulty parameters range value, \code{c(min.beta,max.beta)}. It applies when the \code{randtype = "uniform"}.}

\item{sdbeta}{Standard deviation which is used to generate item location values using \code{\link[stats:list]{rnorm}} with \code{n = nitem}, \code{mean = 0}, and \code{sd = sdbeta}.It applies when the \code{randtype = "normal"}.}

\item{nitem}{The number of the items in each subgroup.}

\item{alpha}{A vector of the discrimination parameters apply to each items.}

\item{sdlambda}{A vector of the standard deviation to simulate the testlet (local dependency) effect. The effect is added using \code{\link[stats:list]{rnorm}} with \code{n = ntheta}, \code{mean = 0}, and \code{sd = sdlambda}}

\item{ncat}{The number of the response categories}

\item{thGap}{The difference between adjacent threshold.}

\item{ndim}{The number of subgroups (dimensions/testlets) created.}

\item{randtype}{The randomize type. This includes \code{uniform} and \code{normal}.}

\item{dim.members}{The list of item members in each dimension.}
}
\description{
This function generates simulated datasets with different attributes
#'
}
\examples{
# Generate multidimensional dataset which having correlation of 0.2 between the dimensions
correl02_multidim <- generate_data(responseType = "multidim.withcorrel", corLevel = 0.2)

# Generate multidimensional dataset with some items relate to more than one dimension.
withinItem_multidim <- generate_data(responseType = "multidim.within", ndim = 3, dim.members = list(c(1:6,13),c(3,7:12),c(5,13:18)))

#generate dataset which consist of two bundle items with different level of local dependency effect.
testlets_dataset <- generate_data(responseType = "testlets", ndim = 2, sdlambda = c(0,4))

}
>>>>>>> b23ed6f4b437001680862795ffb74d5584ee7d26
