% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ieps}
\alias{ieps}
\title{The Interdisciplinary Education Perception Scale (IEPS) Dataset with 2 Covariates}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 319 rows and 20 columns.
}
\usage{
data(ieps)
}
\description{
The pre-processed version of the IEPS dataset in (Vaughan, 2019).
}
\references{
Vaughan, B. (2019). Measurement Properties of the Interdisciplinary Education Perception Scale in an Australian Allied Health Student Cohort. Health Professions Education, 5(3), 275–280. https://doi.org/10.1016/j.hpe.2018.07.005
}
\keyword{datasets}
