% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{inhomogeneous_dataset}
\alias{inhomogeneous_dataset}
\title{inhomogeneous_dataset}
\description{
An artificial dataset which simulates three subscales with different predictability level (discrimination parameters).
}
