% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{inhomogeneous_rasch_dataset}
\alias{inhomogeneous_rasch_dataset}
\title{inhomogeneous_rasch_dataset}
\description{
An artificial dataset of dichotomous responses which simulates three subscales with different predictability level (discrimination parameters).
}
