% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit.R
\name{resid_corr}
\alias{resid_corr}
\title{Residual Correlation}
\usage{
resid_corr(objFit)
}
\arguments{
\item{objFit}{object of class "fit", the output of \code{fitStats()}.}
}
\value{
\item{ld_correl}{  Correlation matrix of the standradized residual.}
\item{ld_mean}{  The mean of the correlation.}
\item{ld_lowertri}{  The lower triangle of the correlation matrix.}
}
\description{
Compute the correlation of the standardized residual to check the local dependency status
}
