% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msqol}
\alias{msqol}
\title{The Multiple Sclerosis Quality of Life (MSQOL)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 473 rows and 55 columns.
}
\usage{
data(msqol)
}
\description{
The pre-processed version of the MSQOL dataset in (Rosato et al., 2016).
}
\references{
Rosato, R., Testa, S., Bertolotto, A., Confalonieri, P., Patti, F., Lugaresi, A., … Solari, A. (2016). Development of a short version of MSQOL- 54 using factor analysis and item response theory. PLoS ONE, 11(4), 1–15. https://doi.org/10.1371/journal.pone.0153466#'
}
\keyword{datasets}
