% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{octq}
\alias{octq}
\title{The Osteopathy Clinical Teaching Questionnaire (OCTQ) Dataset}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 399 rows and 37 columns.
}
\usage{
data(octq)
}
\description{
The pre-processed version of the OCTQ dataset in (Vaughan, 2018).
}
\references{
Vaughan, B. (2018). Exploring the measurement properties of the osteopathy clinical teaching questionnaire using Rasch analysis. Chiropractic and Manual Therapies, 26(1), 1–16. https://doi.org/10.1186/s12998-018-0182-2#'
}
\keyword{datasets}
