% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_function.R
\name{par_map}
\alias{par_map}
\title{Estimated parameters extractor}
\usage{
par_map(nlmPar, estPar_arr, estLength_array, fixValue, fixed_par,
  fixLength_arr)
}
\arguments{
\item{nlmPar}{A vector of the estimated parameters values.}

\item{estPar_arr}{A vector of the types of the estimated parameters.}

\item{estLength_array}{A vector of the lengths of each types of the estimated parameters.}

\item{fixValue}{A vector of the fixed parameters values.}

\item{fixed_par}{A vector of the types of the fixed parameters.}

\item{fixLength_arr}{A vector of the lengths of each types of the fixed parameters.}
}
\value{
output A list of the extracted estimated parameter values.
}
\description{
Mapping the estimated parameters to each variable, i.e., theta, beta, gamma, and delta.
}
