% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pcm_dif}
\alias{pcm_dif}
\alias{summary.pcm_dif}
\alias{print.pcm_dif}
\title{Estimation of The Partial Credit Model with DIF}
\usage{
pcm_dif(X, groups_map, isHessian = TRUE)

\method{summary}{pcm_dif}(obj, par = c(), th_dif = NULL)

\method{print}{pcm_dif}(obj, par = c())
}
\arguments{
\item{X}{Input dataset as matrix or data frame with ordinal responses (starting from 0);
rows represent individuals, columns represent items.}

\item{isHessian}{a logical parameter setting whether or not the Hessian matrix is needed to be computed.
The Hessian matrix is needed to compute the separation reliability.}

\item{obj}{The object of class \code{'pcm'}.}

\item{par}{The parameter that are wanted to be summarized.}

\item{th_dif}{The threshold of a recognizable estimated DIF parameters.}

\item{dif_map}{Binary matrix. Respondents membership to DIF groups; rows represent individuals, column represent group partitions.}
}
\value{
\item{X}{   The dataset that is used for estimation.}
\item{mt_vek}{   A vector of the highest response category as many as the number of items.}
\item{real_vek}{   A vector of the presence of thresholds in the items-thresholds order:
\code{1} if the threshold is present and \code{NA} if unavailable.}
\item{itemName}{   The vector of names of items (columns) in the dataset.}
\item{loglik}{   The log likelihood of the estimation.}
\item{hessian}{   The hessian matrix. Only when the \code{isHessian = TRUE}.}
\item{beta}{   A vector of the difficulty parameter of each categories of items (thresholds).}
\item{theta}{   A vector of the ability parameters of each individuals.}
}
\description{
This function computes the parameter estimates of a partial credit model with DIF for dichotomous and polytomous responses
by using penalized JML estimation.
}
\examples{
res <- pcm_dif(polydif_inh_dset, groups_map = c(rep(1,245),rep(0,245)))
res
summary(res)

}
