% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poly_md_dset}
\alias{poly_md_dset}
\title{Uncorrelated Multidimensional Polytomous Data Set}
\format{
An object of class \code{data.frame} with 301 rows and 18 columns.
}
\usage{
data(poly_md_dset)
}
\description{
Data set with polytomous responses (five categories) containing
three subsets which represent different uncorrelated dimensions.
}
\keyword{datasets}
