% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polydif_inh_dset}
\alias{polydif_inh_dset}
\title{The Inhomogenous Polytomous Dataset containing DIF items}
\format{
An object of class \code{data.frame} with 490 rows and 20 columns.
}
\usage{
data(polydif_inh_dset)
}
\description{
The artificial data set of a polytomous responses (five categories) which
contains three subsets with different discrimination values and two DIF-items.
}
\keyword{datasets}
