% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_search.R
\name{stepwise_search}
\alias{stepwise_search}
\alias{backward_search}
\alias{summary.search}
\alias{plot.search}
\title{Stepwise Selection Search}
\usage{
stepwise_search(X, criterion = c("ipoqll"), incl_set = c(),
  groups_map = c(), cores = NULL, optz_tuner_iq = c(),
  optz_tuner_oq = c(), isTracked = TRUE, isContinued = FALSE,
  prevData = c(), isLegacy = FALSE, fileOutput = FALSE,
  tempFile = "temp_stepSearch.RData", isConvert = FALSE)

backward_search(X, criterion = c("ipoqll"), incl_set = c(),
  groups_map = c(), cores = NULL, optz_tuner_iq = c(),
  optz_tuner_oq = c(), isTracked = TRUE, isContinued = FALSE,
  prevData = c(), isLegacy = FALSE, fileOutput = FALSE,
  tempFile = "temp_backSearch.RData", isConvert = FALSE)

\method{summary}{search}(obj)

\method{plot}{search}(obj, fileOutput = FALSE, type = "l",
  xlab = NULL, ylab = NULL, xlim = c(), ylim = c(),
  use.name = FALSE, remOrdered = TRUE, isMax = TRUE, ...)
}
\arguments{
\item{X}{A matrix or data.frame of the observed responses (ordinal or binary response).}

\item{criterion}{The criterion that should be used. The default is ipoqll.}

\item{incl_set}{A vector of initial items in the included set to start the search. The default is to start with full items.}

\item{groups_map}{A matrix or vector to map the subject to the DIFs groups.}

\item{cores}{An integer value of number of cores should be used for computation. The default is 2.}

\item{optz_tuner_iq}{The optimisation setting of the included set. See \code{\link[stats:optim]{stats::optim()}} \code{control} parameter.}

\item{optz_tuner_oq}{The optimisation setting of the excluded set. See \code{\link[stats:optim]{stats::optim()}} \code{control} parameter.}

\item{isTracked}{A logical value whether the progress need to be tracked or not.}

\item{isContinued}{A logical value whether this search is continuing another unfinished search.}

\item{prevData}{The filename of the temporary .RData file of the unfinished search.}

\item{isLegacy}{A logical value whether using the previous step estimated parameters as the initial parameter values for the upcoming search step.}

\item{fileOutput}{The filename if it is wished to save the output results in file (.RData and .csv) and FALSE if not.}

\item{tempFile}{The filename of the temporary file to track the search progress. The default is \code{"temp_stepSearch.RData"} which also automatically produces \code{"temp_stepSearch.csv"}.}

\item{isConvert}{A logical value whether it is wanted to recompute the score of the search results using IPOQ-LL-DIF criterion.}
}
\value{
A matrix of the itemsets that obtain the highest scores for each number of items in the included set.
}
\description{
To search itemset that give maximum value of the criterion
}
\details{
To search the itemset that give the maximum score.
}
\examples{
search_res <- backward_search(poly_inh_dset,criterion = "ipoqll", incl_set = c(1:18), cores = 2, isLegacy = TRUE)
}
