% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shortDIF}
\alias{shortDIF}
\title{A Shorter Polytomous Dataset with DIF}
\format{
An object of class \code{data.frame} with 100 rows and 4 columns.
}
\usage{
data(shortDIF)
}
\description{
The artificial dataset of a polytomous responses (three categories) which
contains three non-DIF items and a DIF item.
}
\keyword{datasets}
