% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shortDSF}
\alias{shortDSF}
\title{A Shorter Polytomous Dataset with DSF}
\format{
An object of class \code{data.frame} with 400 rows and 6 columns.
}
\usage{
data(shortDSF)
}
\description{
The artificial dataset of a polytomous responses (four categories) which
contains four non-DSF items and two DSF items.
}
\keyword{datasets}
