% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testlets_dataset}
\alias{testlets_dataset}
\title{Multi-testlets Polytomous Data Set}
\format{
An object of class \code{data.frame} with 301 rows and 12 columns.
}
\usage{
data(testlets_dataset)
}
\description{
Generate data set which consist of two bundle items with different level of
local dependency effect.
}
\keyword{datasets}
