% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{uncorrel_rasch_multidim}
\alias{uncorrel_rasch_multidim}
\title{uncorrel_rasch_multidim}
\description{
An artificial dataset of dichotomous responses which simulates three uncorrelated subscales from different dimensions.
}
