% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidimensionality.R
\name{check.unidim}
\alias{check.unidim}
\title{Unidimensionality Check}
\usage{
check.unidim(x, is.polychor = TRUE)
}
\arguments{
\item{x}{The dataset of responses.}

\item{is.polychor}{A boolean parameter to set whether the dataset is categorical or not.}
}
\value{
A list of the CFA output and the some of the goodness-of-fit indices (i.e., cfi, tli, rmsea, and srmr)
}
\description{
This function checks the unidimensionality status using the confirmatory factor analysis.
}
