% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{withinItem_multidim}
\alias{withinItem_multidim}
\title{Within-item Multidimensional Polytomous Data Set}
\format{
An object of class \code{data.frame} with 301 rows and 18 columns.
}
\usage{
data(withinItem_multidim)
}
\description{
Generate multidimensional dataset with some items relate to more than one
dimension.
}
\keyword{datasets}
