% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoRasch_options.R
\name{autoRaschOptions}
\alias{autoRaschOptions}
\title{autoRasch Options}
\usage{
autoRaschOptions(x = NULL)
}
\arguments{
\item{x}{\emph{Character}. A name of single parameter setting that is wanted to be shown.}
}
\value{
\item{fixed_par}{   A vector of parameter types that are set to be fix. It means that these parameters are not estimated.}
\item{fixed_theta}{   A vector of \code{theta} values when \code{theta} are set to be fix at the \code{fixed_par}. If it is not set, it will be set to zero.}
\item{fixed_beta}{   A vector of \code{theta} values when \code{beta} are set to be fix at the \code{fixed_par}. If it is not set, it will be set to zero.}
\item{fixed_gamma}{   A vector of \code{gamma} (natural logarithm of discrimination parameters, \eqn{\alpha = exp(\gamma)}) values when \code{gamma} are set to be fix at the \code{fixed_par}. If it is not set, it will be set to zero.}
\item{isPenalized_theta}{   It is a logical parameter whether, in the estimation procedure, \code{theta} is penalized or not.}
\item{isPenalized_gamma}{   It is a logical parameter whether, in the estimation procedure, \code{gamma} is penalized or not.}
\item{groups_map}{   A matrix \eqn{n x f} to map the subject into DIF groups, where \eqn{n} is number of subjects and \eqn{f} is number of focal groups.}
\item{optz_tuner}{   A list of optimization function settings. For complete settings can be seen in \code{\link[stats:optim]{stats::optim()}}.}
\item{lambda_theta}{   An integer value to set the regularization parameter to the \code{theta}.}
\item{lambda_in}{   An integer value to set the regularization parameter to the \code{gamma} in the included itemset.}
\item{lambda_out}{   An integer value to set the regularization parameter to the \code{gamma} in the excluded itemset.}
\item{isHessian}{   It is a logical parameter whether, in the estimation procedure, need to return the Hessian matrix or not.}
\item{isTracked}{   It is a logical parameter whether need to track the process or not.}
}
\description{
Show the default settings used by the functions in \strong{autoRasch} package. These settings can be changed by passing values through the "..." argument.
}
\examples{
autoRaschOptions()
autoRaschOptions(x = "isTracked")

}
