% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_score.R
\name{compute_score}
\alias{compute_score}
\alias{compute_scores}
\title{Compute the In-plus-out-of-questionnaire log likelihood (IPOQ-LL)}
\usage{
compute_score(X, incl_set, type = c("ipoqll", "ipoqlldif"),
  groups_map = c(), init_par_iq = c(), init_par_oq = c(),
  optz_tuner_iq = c(), optz_tuner_oq = c(), isTracked = FALSE)

compute_scores(X, itemsets, type = c("ipoqll", "ipoqlldif"),
  step_direct = c("fixed", "forward", "backward"), groups_map = c(),
  init_par_iq = c(), init_par_oq = c(), optz_tuner_iq = c(),
  optz_tuner_oq = c(), cores = NULL)
}
\arguments{
\item{X}{A matrix or data.frame of the observed responses (ordinal or binary response).}

\item{incl_set}{A vector of the items (columns) number in the data.frame X that are included in the included set.}

\item{type}{The type of the score. \code{ipoqll} if we ignore the presence of the DIF and \code{ipoqlldif} if we want to consider the DIF effect.}

\item{groups_map}{Matrix to map the respondents to the DIF groups.}

\item{init_par_iq}{Initial values of the parameters in the included set before the estimation begin.}

\item{init_par_oq}{Initial values of the parameters in the excluded set before the estimation begin.}

\item{optz_tuner_iq}{The optimisation setting of the included set. See \code{\link[stats:optim]{stats::optim()}} \code{control} parameter.}

\item{optz_tuner_oq}{The optimisation setting of the excluded set. See \code{\link[stats:optim]{stats::optim()}} \code{control} parameter.}

\item{isTracked}{A logical value whether the computation will be tracked or not.}

\item{cores}{Number of cores that is used in the paralellization.}

\item{incl_sets}{A matrix as a results of a \code{rbind} of \code{incl_set}.}
}
\value{
\code{compute_score} will return a vector which contains in-questionnaire log likelihood (iqll), out-of-questionnaire log likelihood(oqll),
ipoqll, included set's items' number of the given initial survey, the estimated parameters of the included set,
and the estimated parameters of the excluded set, respectively.

\code{compute_scores} will return a matrix as a result of the \code{rbind} operation of the \code{compute_score}'s result.
}
\description{
\code{compute_score} computes the the IPOQ-LL score of an instrument (included set) of the given initial survey.
While \code{compute_scores} computes the IPOQ-LL score of many (more than one) instruments (included sets) of
the given initial survey simultanously.
}
