% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/correl02_multidim.R
\docType{data}
\name{correl02_multidim}
\alias{correl02_multidim}
\title{The Multidimensional Polytomous Dataset with 0.2 Correlation}
\format{An object of class \code{data.frame} with 301 rows and 12 columns.}
\usage{
data(correl02_multidim)
}
\description{
The artifical dataset of a polytomous responses (five categories) which contains of two subsets  which represent different dimensions.
Both dimensions are having 0.2 correlation to each other. To reproduce this dataset: \cr
\code{correl02_multidim <- generate_data(responseType = "multidim.withcorrel", corLevel = 0.2)}
\cr\cr will lead to similar but not the same dataset, due to the randomization.
}
\keyword{datasets}
=======
% Please edit documentation in R/data.R
\name{correl02_multidim}
\alias{correl02_multidim}
\title{correl02_multidim}
\description{
An artificial dataset which simulates two subscales from different dimensions that having correlation of 0.2.
}
>>>>>>> 7bda44cf6ff72132fa57077ea53e1ef9d6063ea5
