% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl04_multidim.R
\docType{data}
\name{correl04_multidim}
\alias{correl04_multidim}
\title{The Multidimensional Polytomous Dataset with 0.4 Correlation}
\format{An object of class \code{data.frame} with 301 rows and 12 columns.}
\usage{
data(correl04_multidim)
}
\description{
The artifical dataset of a polytomous responses (five categories) which contains of two subsets  which represent different dimensions.
Both dimensions are having 0.4 correlation to each other. To reproduce this dataset: \cr
\code{correl04_multidim <- generate_data(responseType = "multidim.withcorrel", corLevel = 0.4)}
\cr\cr will lead to similar but not the same dataset, due to the randomization.
}
\keyword{datasets}
