% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl05_multidim.R
\docType{data}
\name{correl05_multidim}
\alias{correl05_multidim}
\title{The Multidimensional Polytomous Dataset with 0.5 Correlation}
\format{
An object of class \code{data.frame} with 301 rows and 12 columns.
}
\usage{
data(correl05_multidim)
}
\description{
The artifical dataset of a polytomous responses (five categories) which contains of two subsets  which represent different dimensions.
Both dimensions are having 0.5 correlation to each other. To reproduce this dataset: \cr
\code{correl05_multidim <- generate_data(responseType = "multidim.withcorrel", corLevel = 0.5)}
\cr\cr will lead to similar but not the same dataset, due to the randomization.
}
\keyword{datasets}
