% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicho_md_dset.R
\docType{data}
\name{dicho_md_dset}
\alias{dicho_md_dset}
\title{The Uncorrelated Multidimensional Dichotomous Dataset}
\format{An object of class \code{data.frame} with 301 rows and 18 columns.}
\usage{
data(dicho_md_dset)
}
\description{
The artifical dataset of a binary type responses which contains of three subsets which represent different dimensions.
The three dimensions are uncorrelated. To reproduce this dataset: \cr\cr
\code{dicho_md_dset <- generate_data(responseType = "multidim.nocorrel", ncat = 2)}
\cr\cr will lead to similar but not the same dataset, due to the randomization.
}
\keyword{datasets}
