% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit.R
\name{fitStats}
\alias{fitStats}
\alias{summary.fit}
\alias{plot.fit}
\title{FIt statistics}
\usage{
fitStats(obj, isAlpha = TRUE, isTraced = FALSE)

\method{summary}{fit}(obj)

\method{plot}{fit}(obj, plotx = "alpha", ploty = "outfit",
  type = "n", xlab = NULL, ylab = NULL, use.name = FALSE,
  fileOutput = TRUE, ...)
}
\description{
FIt statistics
}
\examples{
res <- pcm(pcm_data)
fit_res <- fitStats(res)
summary(fit_res)
plot(fit_res, plot.x = "gamma", plot.y = "outfit")

}
