% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit.R
\name{fitStats}
\alias{fitStats}
\alias{summary.fit}
\alias{plot.fit}
\title{FIt statistics}
\usage{
fitStats(obj, isAlpha = TRUE, isTraced = FALSE)

\method{summary}{fit}(obj)

\method{plot}{fit}(obj, type = "n", plotx, ploty, xlab = NULL,
  ylab = NULL, use.name = FALSE, ...)
}
\description{
FIt statistics
}
