% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gpcm}
\alias{gpcm}
\alias{summary.gpcm}
\alias{print.gpcm}
\title{Estimation of The Generalized Partial Credit Model}
\usage{
gpcm(X, isHessian = TRUE)

\method{summary}{gpcm}(obj, par = c())

\method{print}{gpcm}(obj, par = c())
}
\arguments{
\item{X}{Input dataset as matrix or data frame with ordinal responses (starting from 0);
rows represent individuals, columns represent items.}

\item{isHessian}{a logical parameter setting whether or not the Hessian matrix is needed to be computed.
The Hessian matrix is needed to compute the separation reliability.}

\item{obj}{The object of class \code{'pcm'}.}

\item{par}{The parameter that are wanted to be summarized.}
}
\value{
\item{X}{   The dataset that is used for estimation.}
\item{mt_vek}{   A vector of the highest response category as many as the number of items.}
\item{real_vek}{   A vector of the presence of thresholds in the items-thresholds order: \code{1} if the threshold is present and \code{NA} if unavailable.}
\item{itemName}{   The vector of names of items (columns) in the dataset.}
\item{loglik}{   The log likelihood of the estimation.}
\item{hessian}{   The hessian matrix. Only when the \code{isHessian = TRUE}.}
\item{gamma}{   A vector of the natural logarithm of discrimination parameters of each items.}
\item{beta}{   A vector of the difficulty parameter of each items' categories (thresholds).}
\item{theta}{   A vector of the ability parameters of each individuals.}
}
\description{
This function computes the parameter estimates of a generalized partial credit model for polytomous responses
by using penalized JML estimation.
}
\details{
In the discrimination parameters estimation, instead of estimating the discrimination parameters,
we are estimating the natural logarithm of the parameters to avoid negative values, \eqn{\alpha = exp(\gamma)}.
}
\examples{
res <- gpcm(poly_inh_dset)
res
summary(res)

}
