% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit.R
\name{residCor}
\alias{residCor}
\alias{corResid}
\alias{summary.ld}
\title{Residual Correlation}
\usage{
residCor(objFit)

corResid(objFit)

\method{summary}{ld}(object, ...)
}
\arguments{
\item{objFit}{object of class "fit", the output of \code{fitStats()}.}

\item{object}{The object of class \code{'ld'}.}

\item{...}{Further arguments to be passed.}
}
\value{
\item{ld_correl}{  Correlation matrix of the standradized residual.}
\item{ld_mean}{  The mean of the correlation.}
\item{ld_lowertri}{  The lower triangle of the correlation matrix.}
}
\description{
Compute the correlation of the standardized residual to check the local dependency status
}
