% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pcm}
\alias{pcm}
\alias{summary.pcm}
\alias{print.pcm}
\title{Estimation of The Partial Credit Model (PCM)}
\usage{
pcm(X, isHessian = TRUE)

\method{summary}{pcm}(obj, par = c())

\method{print}{pcm}(obj, par = c())
}
\arguments{
\item{X}{Input dataset as matrix or data frame with ordinal responses (starting from 0);
rows represent individuals, columns represent items.}

\item{isHessian}{a logical parameter setting whether or not the Hessian matrix is needed to be computed.
The Hessian matrix is needed to compute the separation reliability.}

\item{obj}{The object of class \code{'pcm'}.}

\item{par}{The parameter that are wanted to be summarized.}
}
\value{
\item{X}{   The dataset that is used for estimation.}
\item{mt_vek}{   A vector of the highest response category as many as the number of items.}
\item{real_vek}{   A vector of the presence of thresholds in the items-thresholds order: \code{1} if the threshold is present and \code{NA} if unavailable.}
\item{itemName}{   The vector of names of items (columns) in the dataset.}
\item{loglik}{   The log likelihood of the estimation.}
\item{hessian}{   The hessian matrix. Only when the \code{isHessian = TRUE}.}
\item{beta}{   A vector of the difficulty parameter of each categories of items (thresholds).}
\item{theta}{   A vector of the ability parameters of each individuals.}
}
\description{
This function computes the parameter estimates of a partial credit model for dichotomous and polytomous responses
by using penalized JML estimation.
}
\examples{
res <- pcm(poly_inh_dset)
res
summary(res)

}
