% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcmdif.R
\name{pcm_dif}
\alias{pcm_dif}
\alias{fitStats.pcmdif}
\alias{summary.pcm_dif}
\alias{print.pcm_dif}
\title{Estimation of The Partial Credit Model with DIF}
\usage{
pcm_dif(X, groups_map, isHessian = TRUE)

\method{fitStats}{pcmdif}(obj, isAlpha = TRUE)

\method{summary}{pcm_dif}(obj, par = c(), th_dif = NULL)

\method{print}{pcm_dif}(obj, par = c())
}
\arguments{
\item{X}{Input dataset as matrix or data frame with ordinal responses (starting from 0);
rows represent individuals, columns represent items.}

\item{groups_map}{Binary matrix. Respondents membership to DIF groups; rows represent individuals, column represent group partitions.}

\item{isHessian}{a logical parameter setting whether or not the Hessian matrix is needed to be computed.
The Hessian matrix is needed to compute the separation reliability.}

\item{obj}{The object of class \code{'pcm_dif'}.}

\item{isAlpha}{Boolean value that indicates whether the discrimination parameters is needed to be estimated or not.
The discrimination parameters are estimated using the corresponding models (GPCM or GPCM-DIF).}

\item{par}{The parameter that are wanted to be summarized.}

\item{th_dif}{The threshold of a recognizable estimated DIF parameters.}
}
\value{
\strong{\code{pcm_dif()} will return a \code{\link[base:list]{list}} which contains:}
\item{X}{   The dataset that is used for estimation.}
\item{mt_vek}{   A vector of the highest response category as many as the number of items.}
\item{real_vek}{   A vector of the presence of thresholds in the items-thresholds order:
\code{1} if the threshold is present and \code{NA} if unavailable.}
\item{itemName}{   The vector of names of items (columns) in the dataset.}
\item{loglik}{   The log likelihood of the estimation.}
\item{hessian}{   The hessian matrix. Only when the \code{isHessian = TRUE}.}
\item{beta}{   A vector of the difficulty parameter of each categories of items (thresholds).}
\item{theta}{   A vector of the ability parameters of each individuals.}

\strong{\code{fitStats()} will return a \code{\link[base:list]{list}} which contains:}
\item{alpha}{   A vector of estimated discrimination parameters for each items.}
\emph{i.fit}{   Item fit statistics.}
\itemize{
   \item{i.outfitMSQ}{   A vector of Outfit mean square values for each items.}
   \item{i.infitMSQ}{   A vector of Infit mean square values for each items.}
   \item{i.outfitZ}{   A vector of OutfitZ values for each items.}
   \item{i.infitZ}{   A vector of InfitZ values for each items.}
}
\emph{p.fit}{   Person fit statistics.}
\itemize{
   \item{p.outfitMSQ}{   A vector of Outfit mean square values for each persons.}
   \item{p.infitMSQ}{   A vector of Infit mean square values for each persons.}
   \item{p.outfitZ}{   A vector of OutfitZ values for each persons.}
   \item{p.infitZ}{   A vector of InfitZ values for each persons.}
}
}
\description{
This function computes the parameter estimates of a partial credit model with DIF for dichotomous and polytomous responses
by using penalized JML estimation.

\code{fitStats} compute the fit statistics (e.g., Outfit and Infit) of the PCM-DIF model estimation (items and persons).
}
\examples{
res <- pcm_dif(polydif_inh_dset, groups_map = c(rep(1,245),rep(0,245)))
res
summary(res)

pcmdif_res <- pcm_dif(polydif_inh_dset, groups_map = c(rep(1,245),rep(0,245)))
fit_res <- fitStats(pcmdif_res)
summary(fit_res)
plot(fit_res, plotx = "gamma", ploty = "outfit")

}
\seealso{
\code{\link{pcm}}, \code{\link{pcm_dif}}, \code{\link{gpcm}}, \code{\link{gpcm_dif}}
}
