% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_md_dset.R
\docType{data}
\name{poly_md_dset}
\alias{poly_md_dset}
\title{The Uncorrelated Multidimensional Polytomous Dataset}
\format{An object of class \code{data.frame} with 301 rows and 18 columns.}
\usage{
data(poly_md_dset)
}
\description{
The artifical dataset of a polytomous responses (five categories) which contains of three subsets  which represent different dimensions.
The three dimensions are uncorrelated. To reproduce this dataset: \cr
\code{poly_md_dset <- generate_data(responseType = "multidim.nocorrel", ncat = 5)}
\cr\cr will lead to similar but not the same dataset, due to the randomization.
}
\keyword{datasets}
