% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testlets_dataset.R
\docType{data}
\name{testlets_dataset}
\alias{testlets_dataset}
\title{The Multi-testlets Polytomous Dataset}
\format{An object of class \code{data.frame} with 301 rows and 12 columns.}
\usage{
data(testlets_dataset)
}
\description{
The artifical dataset of a polytomous responses (five categories) which contains of two subsets which represent different testlets.
Both testlets are having difference variance effects. To reproduce this dataset: \cr
\code{testlets_dataset <- generate_data(responseType = "testlets", ndim = 2, sdlambda = c(0,4))}
\cr\cr will lead to similar but not the same dataset, due to the randomization.
}
\keyword{datasets}
