% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{withinItem_multidim}
\alias{withinItem_multidim}
\title{withinItem_multidim}
\description{
An artificial dataset which consist of three subscales (dimensions) with some of items relate to more than one subscale (dimension).
}
