% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinItem_multidim.R
\docType{data}
\name{withinItem_multidim}
\alias{withinItem_multidim}
\title{The Within-item Multidimensional Polytomous Dataset}
\format{
An object of class \code{data.frame} with 301 rows and 18 columns.
}
\usage{
data(withinItem_multidim)
}
\description{
The artifical dataset of a polytomous responses (five categories) which contains of three subsets which represent different dimensions.
The three dimensions are uncorrelated, however, some of items relate to more than one dimension. To reproduce this dataset: \cr
\code{withinItem_multidim <- generate_data(responseType = "multidim.within", ndim = 3, dim.members = list(c(1:6,13),c(3,7:12),c(5,13:18)))}
\cr\cr will lead to similar but not the same dataset, due to the randomization.
}
\keyword{datasets}
